# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Customer API
  description: API para gestionar clientes y sus productos asociados
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:
  /api/customers:
    get:
      summary: Listar todos los clientes
      operationId: getAllCustomers
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
              examples:
                sample:
                  summary: ejemplo lista
                  value:
                    - id: "1"
                      customerType: "PERSONAL"
                      firstName: "Juan"
                      lastName: "Perez"
    post:
      summary: Crear un cliente
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
            example:
              customerType: "PERSONAL"
              firstName: "Ana"
              lastName: "Gomez"
              dni: "12345678"
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
              example:
                id: "generated-id-123"
                customerType: "PERSONAL"
                firstName: "Ana"
                lastName: "Gomez"
                dni: "12345678"

  /api/customers/{id}:
    get:
      summary: Obtener cliente por ID
      operationId: getCustomerById
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
              example:
                id: "1"
                customerType: "PERSONAL"
                firstName: "Juan"
                lastName: "Perez"
    put:
      summary: Actualizar cliente
      operationId: updateCustomer
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    delete:
      summary: Eliminar cliente
      operationId: deleteCustomer
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Cliente eliminado
  /api/customers/{id}/products:
    get:
      summary: Listar productos por id de un cliente
      operationId: getCustomerProducts
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Lista de productos del cliente por id
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      summary: Agrega el id de un producto a un cliente
      operationId: addProductToCustomer
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductIdRequest'
      responses:
        '204':
          description: Producto agregado al cliente
  /api/customers/{id}/products/{productId}:
    delete:
      summary: Elimina un producto ID de un cliente
      operationId: removeProductFromCustomer
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: path
          name: productId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Producto eliminado del cliente
components:
  schemas:
    # Request para crear/actualizar (sin id)
    CustomerRequest:
      type: object
      required:
        - customerType
        - firstName
      properties:
        customerType:
          type: string
          enum: [PERSONAL, EMPRESA]
        firstName:
          type: string
        lastName:
          type: string
        businessName:
          type: string
        dni:
          type: string
        ruc:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        productIds:
          type: array
          items:
            type: string

    # Response que incluye id (lo que devuelve la API)
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
        customerType:
          type: string
          enum: [PERSONAL, EMPRESA]
        firstName:
          type: string
        lastName:
          type: string
        businessName:
          type: string
        dni:
          type: string
        ruc:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        productIds:
          type: array
          items:
            type: string
    ProductIdRequest:
      type: object
      required:
        - productId
      properties:
        productId:
          type: string